revise type: int -> size_t

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: athonda <athonda@student.42singapore.sg    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/14 07:46:03 by athonda           #+#    #+#             */
-/*   Updated: 2024/05/17 14:23:46 by athonda          ###   ########.fr       */
+/*   Updated: 2024/05/24 11:59:25 by athonda          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
